backend/requirements.txt
fastapi==0.116.0
uvicorn==0.34.2
pydantic==2.11.4
PyYAML==6.0.2
requests==2.32.3
openai
python-dotenv
